Update gift_ideas.py
Trying to add items. Might need to change what's being passed to the editing functions so that the dictionary is updated. Maybe put that in the menu selection block (i.e., after returning an updated list, in the if/elif block of the menu selection edit the key/value pair for the dictionary).